pyproject.toml
[tool.poetry]
name = "pass-ssh-agent"
version = "0.1.0"
description = "SSH Agent backed by passwordstore"
authors = ["Brett Jones <noreply@blockloop.io>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pyright = "^1.1.305"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
